+++ src/atlas/util/io/Gmsh.cc
@@ -111,7 +111,7 @@ void write_field_nodes(const Metadata& gmsh_options, const functionspace::NodeCo
   bool gather( gmsh_options.get<bool>("gather") );
   bool binary( !gmsh_options.get<bool>("ascii") );
   int nlev  = field.levels();
-  int ndata = std::min(function_space.nb_nodes(),field.shape(0));
+  size_t ndata = std::min(function_space.nb_nodes(),field.shape(0));
   int nvars = field.stride(0)/nlev;
   array::ArrayView<gidx_t,1> gidx ( function_space.nodes().global_index() );
   array::ArrayView<DATATYPE,2> data ( field.data<DATATYPE>(), array::make_shape(field.shape(0),field.stride(0)) );
@@ -142,7 +142,7 @@ void write_field_nodes(const Metadata& gmsh_options, const functionspace::NodeCo
   {
     lev = gmsh_levels;
   }
-  for (int ilev=0; ilev<lev.size(); ++ilev)
+  for (size_t ilev=0; ilev < lev.size(); ++ilev)
   {
     int jlev = lev[ilev];
     if( ( gather && atlas::parallel::mpi::comm().rank() == 0 ) || !gather )
@@ -170,7 +170,7 @@ void write_field_nodes(const Metadata& gmsh_options, const functionspace::NodeCo
         if( nvars == 1)
         {
           double value;
-          for( int n = 0; n < ndata; ++n )
+          for(size_t n = 0; n < ndata; ++n)
           {
             out.write(reinterpret_cast<const char*>(&gidx(n)),sizeof(int));
             value = data(n,jlev*nvars+0);
@@ -180,7 +180,7 @@ void write_field_nodes(const Metadata& gmsh_options, const functionspace::NodeCo
         else if( nvars <= 3 )
         {
           double value[3] = {0,0,0};
-          for( size_t n = 0; n < ndata; ++n )
+          for(size_t n = 0; n < ndata; ++n)
           {
             out.write(reinterpret_cast<const char*>(&gidx(n)),sizeof(int));
             for( int v=0; v<nvars; ++v)
@@ -197,7 +197,7 @@ void write_field_nodes(const Metadata& gmsh_options, const functionspace::NodeCo
           };
           if( nvars == 4 )
           {
-            for( size_t n = 0; n < ndata; ++n )
+            for(size_t n = 0; n < ndata; ++n)
             {
               out.write(reinterpret_cast<const char*>(&gidx(n)),sizeof(int));
               for( int i=0; i<2; ++i )
@@ -212,7 +212,7 @@ void write_field_nodes(const Metadata& gmsh_options, const functionspace::NodeCo
           }
           if( nvars == 9 )
           {
-            for( size_t n = 0; n < ndata; ++n )
+            for(size_t n = 0; n < ndata; ++n)
             {
               out.write(reinterpret_cast<const char*>(&gidx(n)),sizeof(int));
               for( int i=0; i<3; ++i )
@@ -232,7 +232,7 @@ void write_field_nodes(const Metadata& gmsh_options, const functionspace::NodeCo
       {
         if( nvars == 1)
         {
-          for( int n = 0; n < ndata; ++n )
+          for( size_t n = 0; n < ndata; ++n )
           {
             ASSERT( jlev*nvars < data.shape(1) );
             ASSERT( n < gidx.shape(0) );
@@ -335,7 +335,7 @@ void write_field_nodes(
         field_glb->stride(0)));
     }
 
-    int ndata = data.shape(0);
+    size_t ndata = data.shape(0);
 
     std::vector<long> lev;
     std::vector<long> gmsh_levels;
@@ -356,7 +356,7 @@ void write_field_nodes(
 
     if (atlas::parallel::mpi::comm().rank() == 0)
     {
-        for (int ilev = 0; ilev < lev.size(); ++ilev)
+        for (size_t ilev = 0; ilev < lev.size(); ++ilev)
         {
             int jlev = lev[ilev];
             char field_lev[6] = {0, 0, 0, 0, 0, 0};
@@ -389,7 +389,7 @@ void write_field_nodes(
                 if (nvars == 1)
                 {
                     double value;
-                    for (int n = 0; n < ndata; ++n)
+                    for (size_t n = 0; n < ndata; ++n)
                     {
                         out.write(reinterpret_cast<const char*>(n+1),
                                   sizeof(int));
@@ -422,7 +422,7 @@ void write_field_nodes(
                 ASSERT(jlev*nvars <= data.shape(1));
                 if (nvars == 1)
                 {
-                    for (int n = 0; n < ndata; ++n)
+                    for (size_t n = 0; n < ndata; ++n)
                     {
                         ASSERT(n < data.shape(0));
                         out << n+1 << " "
